% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCohortTable.R
\name{collect.cohort_table}
\alias{collect.cohort_table}
\title{To collect a `cohort_table` object.}
\usage{
\method{collect}{cohort_table}(x, ...)
}
\arguments{
\item{x}{`cohort_table` object.}

\item{...}{Not used (for compatibility).}
}
\value{
A data frame with the `cohort_table`
}
\description{
To collect a `cohort_table` object.
}
