% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodelist.R
\name{emptyCodelist}
\alias{emptyCodelist}
\title{Empty `codelist` object.}
\usage{
emptyCodelist()
}
\value{
An empty codelist object.
}
\description{
Empty `codelist` object.
}
\examples{
emptyCodelist()

}
